Moving FocusZone styles to it's own file in FocusZone folder
